+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[Breaking]** Replace `"browser"` with ES5 CommonJS modules by `"module""` with ES6 modules in `package.json`
 - **[Breaking]** Drop library support of ES5, use a shim or transpiler if needed.
 - **[Breaking]** Mark `unorm` and `bson` as optional dependencies.
+- **[Internal]** Convert gulpfile to Typescript
 
 # 0.5.0-alpha.9 (2017-07-14)